Gradle structure clean-up, comments clean-ups
