+++ basic/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id(Plugins.BASE_GRADLE_CONFIG)
+}
+
+dependencies {
+    implementation(project(Modules.CORE))
+}
