--- basic/src/main/java/com/tamimattafi/navigation/basic/BasicNavigator.kt
@@ -1,15 +1,0 @@
-package com.tamimattafi.navigation.basic
-
-import com.tamimattafi.navigation.core.NavigationContract.Navigator
-import com.tamimattafi.navigation.basic.fragments.BasicNavigationFragment
-
-
-/*
-* #DESCRIPTION
-*
-* For documentations visit NavigationContract.Navigator<T> class
-* An instance of this interface will be passed to all of your BasicNavigationFragments
-* All behaviours from NavigationContract.Navigator<T> apply to this interface
-*
-* */
-interface BasicNavigator : Navigator<BasicNavigationFragment>
