+++ build.gradle.kts
@@ -1,0 +1,13 @@
+
+
+allprojects {
+    repositories {
+        google()
+        jcenter()
+        mavenCentral()
+        maven("https://maven.fabric.io/public")
+        maven("https://plugins.gradle.org/m2/")
+        maven("https://jitpack.io")
+    }
+
+}
