+++ buildSrc/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    `java-gradle-plugin`
+    `kotlin-dsl`
+}
+
+repositories {
+    google()
+    jcenter()
+    mavenCentral()
+}
+
+kotlinDslPluginOptions {
+    experimentalWarning.set(false)
+}
+
+object PluginsVersions {
+    const val GRADLE_ANDROID = "3.5.3"
+    const val KOTLIN = "1.3.61"
+}
+
+dependencies {
+    implementation("com.android.tools.build:gradle:${PluginsVersions.GRADLE_ANDROID}")
+    implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:${PluginsVersions.KOTLIN}")
+
+}
