+++ buildSrc/src/main/kotlin/AndroidConfig.kt
@@ -1,0 +1,7 @@
+object AndroidConfig {
+    const val COMPILE_SDK_VERSION = 29
+    const val MIN_SDK_VERSION = 14
+    const val TARGET_SDK_VERSION = 29
+    const val VERSION_CODE = 8
+    const val VERSION_NAME = "1.0.8"
+}
