+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -1,0 +1,11 @@
+object Dependencies {
+    const val KOTLIN = "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${Versions.KOTLIN_VERSION}"
+    const val APPCOMPAT = "androidx.appcompat:appcompat:${Versions.APPCOMPAT_VERSION}"
+    const val KTX = "androidx.core:core-ktx:${Versions.KTX_VERSION}"
+
+    const val DAGGER = "com.google.dagger:dagger:${Versions.DAGGER_VERSION}"
+    const val DAGGER_COMPILLER = "com.google.dagger:dagger-compiler:${Versions.DAGGER_VERSION}"
+    const val DAGGER_ANDROID = "com.google.dagger:dagger-android:${Versions.DAGGER_VERSION}"
+    const val DAGGER_ANDROID_SUPPORT = "com.google.dagger:dagger-android-support:${Versions.DAGGER_VERSION}"
+    const val DAGGER_ANDROID_PROCESSOR = "com.google.dagger:dagger-android-processor:${Versions.DAGGER_VERSION}"
+}
