+++ buildSrc/src/main/kotlin/Modules.kt
@@ -1,0 +1,4 @@
+object Modules {
+
+    const val CORE = ":core"
+}
