+++ buildSrc/src/main/kotlin/Plugins.kt
@@ -1,0 +1,7 @@
+object Plugins {
+    const val ANDROID_LIBRARY = "com.android.library"
+    const val KOTLIN_ANDROID = "kotlin-android"
+    const val KOTLIN_ANDROID_EXTENSIONS = "kotlin-android-extensions"
+    const val KOTLIN_KAPT = "kotlin-kapt"
+    const val BASE_GRADLE_CONFIG = "customplugins.android-base"
+}
