+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,0 +1,6 @@
+object Versions {
+    const val DAGGER_VERSION = "2.24"
+    const val KOTLIN_VERSION = "1.3.61"
+    const val APPCOMPAT_VERSION = "1.1.0"
+    const val KTX_VERSION = "1.2.0"
+}
