+++ buildSrc/src/main/kotlin/customplugins/android-base.gradle.kts
@@ -1,0 +1,44 @@
+package customplugins
+
+plugins {
+    id("com.android.library")
+    id("kotlin-android")
+    id("kotlin-android-extensions")
+}
+
+android {
+    compileSdkVersion(AndroidConfig.COMPILE_SDK_VERSION)
+
+    defaultConfig {
+        minSdkVersion(AndroidConfig.MIN_SDK_VERSION)
+        targetSdkVersion(AndroidConfig.TARGET_SDK_VERSION)
+    }
+
+    sourceSets {
+        getByName("main") {
+            java.srcDir("src/main/kotlin")
+        }
+    }
+
+    configurations {
+        all {
+            listOf(
+                "META-INF/DEPENDENCIES",
+                "META-INF/LICENSE",
+                "META-INF/LICENSE.txt",
+                "META-INF/license.txt",
+                "META-INF/NOTICE",
+                "META-INF/NOTICE.txt",
+                "META-INF/notice.txt",
+                "META-INF/ASL2.0",
+                "META-INF/app_release.kotlin_module"
+            ).forEach { exclude(it) }
+        }
+    }
+}
+
+dependencies {
+    implementation(Dependencies.KOTLIN)
+    implementation(Dependencies.KTX)
+    implementation(Dependencies.APPCOMPAT)
+}
