+++ core/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id(Plugins.BASE_GRADLE_CONFIG)
+}
