+++ core/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.tamimattafi.navigation.core" />
+<manifest package="com.tamimattafi.navigation.core" />
