+++ core/src/main/kotlin/com/tamimattafi/navigation/core/fragments/BaseNavigationFragment.kt
@@ -1,0 +1,28 @@
+package com.tamimattafi.navigation.core.fragments
+
+import android.util.Log
+import com.tamimattafi.navigation.core.NavigationContract.SelectionListener
+import com.tamimattafi.navigation.core.animation.AnimationSet
+
+abstract class BaseNavigationFragment : BaseFragment(), SelectionListener {
+
+    /**
+    * Fragment name that will be saved in back-stack and used for logging, make sure to have a unique name for each fragment type
+    * Calling this.javaClass.name is a good idea if your fragment classes have unique names
+    */
+    abstract var fragmentName: String
+
+    /**
+    * Animation that will be played during each transaction.
+    */
+    open var animationSet: AnimationSet? = AnimationSet.DEFAULT
+
+    /**
+    * Will be called when a navigation to that fragment happens.
+    * be careful from using any view components because they might still be null when this method is called
+    */
+    override fun onSelected() {
+        Log.d(fragmentName, "Selected")
+    }
+
+}
