+++ settings.gradle.kts
@@ -1,0 +1,8 @@
+rootProject.buildFileName = "build.gradle.kts"
+rootProject.name = "NavigationManager"
+
+include (
+    ":dagger",
+    ":basic",
+    ":core"
+)
